# static/swagger.json
{"info": {"description": "API for accessing Congress.gov data", "contact": {"email": "support@example.com"}, "title": "Congress Data API", "version": "1.0.0"}, "paths": {"/api/bills": {"get": {"summary": "Get bills", "description": "Retrieve bills with optional filtering by congress, bill_type, and date range", "parameters": [{"in": "query", "name": "congress", "schema": {"type": "integer"}, "description": "Filter by congress number (e.g., 117)"}, {"in": "query", "name": "bill_type", "schema": {"type": "string"}, "description": "Filter by bill type (e.g., hr, s)"}, {"in": "query", "name": "start_date", "schema": {"type": "string", "format": "date"}, "description": "Filter by update date (start date, format YYYY-MM-DD)"}, {"in": "query", "name": "end_date", "schema": {"type": "string", "format": "date"}, "description": "Filter by update date (end date, format YYYY-MM-DD)"}, {"in": "query", "name": "limit", "schema": {"type": "integer", "default": 20}, "description": "Maximum number of results to return"}], "responses": {"200": {"description": "Successful response", "content": {"application/json": {"schema": {"type": "object", "properties": {"bills": {"type": "array", "items": {"$ref": "#/components/schemas/Bill"}}, "count": {"type": "integer"}, "next_token": {"type": "string"}}}}}}, "400": {"description": "Bad request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "500": {"description": "Server error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}}}}, "/api/committees": {"get": {"summary": "Get committees", "description": "Retrieve committees with optional filtering by congress, chamber, and date range", "parameters": [{"in": "query", "name": "congress", "schema": {"type": "integer"}, "description": "Filter by congress number (e.g., 117)"}, {"in": "query", "name": "chamber", "schema": {"type": "string"}, "description": "Filter by chamber (House, Senate)"}, {"in": "query", "name": "start_date", "schema": {"type": "string", "format": "date"}, "description": "Filter by update date (start date, format YYYY-MM-DD)"}, {"in": "query", "name": "end_date", "schema": {"type": "string", "format": "date"}, "description": "Filter by update date (end date, format YYYY-MM-DD)"}, {"in": "query", "name": "limit", "schema": {"type": "integer", "default": 20}, "description": "Maximum number of results to return"}], "responses": {"200": {"description": "Successful response", "content": {"application/json": {"schema": {"type": "object", "properties": {"committees": {"type": "array", "items": {"$ref": "#/components/schemas/Committee"}}, "count": {"type": "integer"}, "next_token": {"type": "string"}}}}}}, "400": {"description": "Bad request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "500": {"description": "Server error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}}}}, "/api/hearings": {"get": {"summary": "Get hearings", "description": "Retrieve hearings with optional filtering by congress, committee, and date range", "parameters": [{"in": "query", "name": "congress", "schema": {"type": "integer"}, "description": "Filter by congress number (e.g., 117)"}, {"in": "query", "name": "committee", "schema": {"type": "string"}, "description": "Filter by committee system code"}, {"in": "query", "name": "chamber", "schema": {"type": "string"}, "description": "Filter by chamber (House, Senate)"}, {"in": "query", "name": "start_date", "schema": {"type": "string", "format": "date"}, "description": "Filter by hearing date (start date, format YYYY-MM-DD)"}, {"in": "query", "name": "end_date", "schema": {"type": "string", "format": "date"}, "description": "Filter by hearing date (end date, format YYYY-MM-DD)"}, {"in": "query", "name": "limit", "schema": {"type": "integer", "default": 20}, "description": "Maximum number of results to return"}], "responses": {"200": {"description": "Successful response", "content": {"application/json": {"schema": {"type": "object", "properties": {"hearings": {"type": "array", "items": {"$ref": "#/components/schemas/Hearing"}}, "count": {"type": "integer"}, "next_token": {"type": "string"}}}}}}, "400": {"description": "Bad request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "500": {"description": "Server error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}}}}}, "openapi": "3.0.2", "components": {"schemas": {"Error": {"type": "object", "properties": {"error": {"type": "string", "description": "Error message"}, "status": {"type": "integer", "description": "HTTP status code"}}}, "Bill": {"type": "object", "properties": {"id": {"type": "string", "description": "Unique identifier for the bill"}, "type": {"type": "string", "description": "Type of data (always 'bill')"}, "congress": {"type": "integer", "description": "Congress number"}, "update_date": {"type": "string", "format": "date", "description": "Last update date"}, "bill_type": {"type": "string", "description": "Type of bill (hr, s, etc.)"}, "bill_number": {"type": "integer", "description": "Bill number"}, "title": {"type": "string", "description": "Bill title"}, "origin_chamber": {"type": "string", "description": "Chamber where bill originated"}, "latest_action": {"type": "object", "properties": {"text": {"type": "string", "description": "Latest action text"}, "action_date": {"type": "string", "format": "date", "description": "Action date"}}}}}, "Committee": {"type": "object", "properties": {"id": {"type": "string", "description": "Unique identifier for the committee"}, "type": {"type": "string", "description": "Type of data (always 'committee')"}, "congress": {"type": "integer", "description": "Congress number"}, "update_date": {"type": "string", "format": "date", "description": "Last update date"}, "name": {"type": "string", "description": "Committee name"}, "chamber": {"type": "string", "description": "Chamber (House/Senate)"}, "committee_type": {"type": "string", "description": "Committee type (standing, etc.)"}, "system_code": {"type": "string", "description": "Committee system code"}}}, "Hearing": {"type": "object", "properties": {"id": {"type": "string", "description": "Unique identifier for the hearing"}, "type": {"type": "string", "description": "Type of data (always 'hearing')"}, "congress": {"type": "integer", "description": "Congress number"}, "update_date": {"type": "string", "format": "date", "description": "Last update date"}, "chamber": {"type": "string", "description": "Chamber (House/Senate)"}, "date": {"type": "string", "format": "date", "description": "Hearing date"}, "time": {"type": "string", "description": "Hearing time"}, "location": {"type": "string", "description": "Hearing location"}, "title": {"type": "string", "description": "Hearing title"}, "committee": {"type": "object", "description": "Committee holding the hearing"}}}}}}